update per feedback
